Move local query mappers to separate files
